Update Extended Euclid’s Algorithm.cpp
We can solve also this problem using the formula gcd(a, b) = gcd(b, a mod b).
Suppose that we have already solved the problem for gcd(b, a mod b), and we know
values x' and y' for which
bx' + (a mod b)y' = gcd(a, b).
Then, since a mod b = a − floor(a/b) · b,
bx' + (a − floor(a/b) · b)y' = gcd(a, b),
which equals
ay' + b(x' − floor(a/b) · y') = gcd(a, b).
Thus, we can choose x = y' and y = x' − floor(a/b) · y'. Using this idea, the following function returns a tuple (x, y, gcd(a, b)) that satisfies the equation.